added confirm()-method for a dialog

+++ frontend/src/app/components/bonuses/bonuses.component.ts
@@ -42,8 +42,10 @@ export class BonusesComponent implements OnInit {
     }
 
     delete(bonus: Bonus): void {
-        this.bonuses = this.bonuses.filter((b): boolean => b !== bonus);
-        this.bonusService.deleteBonus(bonus.id)
-            .subscribe();
+        if (confirm('Are you sure to delete this bonus?')) {
+            this.bonuses = this.bonuses.filter((b): boolean => b !== bonus);
+            this.bonusService.deleteBonus(bonus.id)
+                .subscribe();
+        }
     }
 }
